#TODO
• Better Description
• Better Code
• Better Emojis & Formatting

+++ plugins/_ultroid.py
@@ -4,9 +4,8 @@
 # This file is a part of < https://github.com/TeamUltroid/Ultroid/ >
 # PLease read the GNU Affero General Public License in
 # <https://www.github.com/TeamUltroid/Ultroid/blob/main/LICENSE/>.
-from telethon.errors import ChatSendInlineForbiddenError
-from telethon.errors.rpcerrorlist import BotMethodInvalidError as bmi
 
+from telethon.errors import ChatSendInlineForbiddenError
 from . import *
 
 REPOMSG = """
@@ -16,13 +15,18 @@ REPOMSG = """
 • Support - @UltroidSupport
 """
 
+RP_BUTTONS = [[Button.inline("Repo", "https://github.com/TeamUltroid/Ultroid"),
+              Button.inline("Addons","https://github.com/TeamUltroid/UltroidAddons")],
+              [Button.inline("Support Group","t.me/ultroidsupport")]]
 
 @ultroid_cmd(pattern="repo$", type=["official", "manager"], ignore_dualmode=True)
 async def repify(e):
+    if e.client._bot:
+        return await e.reply("**Ultroid Userbot**", buttons=RP_BUTTONS)
     try:
         q = await e.client.inline_query(asst.me.username, "repo")
         await q[0].click(e.chat_id)
         if e.out:
             await e.delete()
-    except (ChatSendInlineForbiddenError, bmi):
+    except (ChatSendInlineForbiddenError):
         await eor(e, REPOMSG)
